% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busdater-deprecated.R
\name{busdater-deprecated}
\alias{busdater-deprecated}
\title{Deprecated functions in busdater}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{FY}}
 \item \code{\link{period_boundaries}}
}
}
