% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_dates.R
\docType{package}
\name{busdater-package}
\alias{busdater}
\alias{busdater-package}
\title{busdater: Standard Date Calculations for Business}
\description{
Get a current financial year, start of current
    month, End of current month, start of financial year and end of it.
    Allow for offset from the date.
}
\details{
Working with business dates can be cumbersome and error prone.
This package aims to make it easier, by providing functionality,
to get financial year, calendar year and month calculations.
It returns start and end of the business periods.

\describe{
  \item{\code{\link{get_fy}}}{Get financial year, possibly with an offset period}
  \item{\code{\link{FY}}}{Deprecated in favour of get_fy()}
  \item{\code{\link{get_boundary}}}{Get date's business period boundary.}
  \item{\code{\link{period_boundaries}}}{Deprecated in favour of get_boundary()}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mickmioduszewski.github.io/busdater/}
  \item \url{https://github.com/mickmioduszewski/busdater/}
  \item Report bugs at \url{https://github.com/mickmioduszewski/busdater/issues}
}

}
\author{
\strong{Maintainer}: Mick Mioduszewski \email{mick@mioduszewski.net}

}
\keyword{internal}
