% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_dates.R
\docType{package}
\name{busdater-package}
\alias{busdater}
\alias{busdater-package}
\title{busdater: Standard Date Calculations for Business}
\description{
Get a current financial year, start of current month, 
    End of current month, start of financial year and end of it.
    Allow for offset from the date.
}
\details{
Working with business dates can be cumbersome and error prone.
This package aims to make it easier, by providing functionality,
to get financial year, calendar year and month calculations.
It returns start and end of the business periods.

\describe{
  \item{\code{\link{FY}}}{Get financial year, possibly with an offset priod}
  \item{\code{\link{period_boundaries}}}{Get date's business period boundary.}
}
}
\author{
\strong{Maintainer}: Mick Mioduszewski \email{mick@mioduszewski.net}

}
\keyword{internal}
